+++ app/apps/web/src/main.tsx
@@ -5,6 +5,7 @@ import "@fontsource/barlow-condensed/700.css";
 import "@fontsource-variable/inter";
 import "@fontsource-variable/noto-sans-sc";
 import { App } from "./App";
+import { I18nProvider } from "./i18n/I18nProvider";
 import "./styles.css";
 
 const root = document.getElementById("root");
@@ -15,7 +16,9 @@ if (!root) {
 
 createRoot(root).render(
   <React.StrictMode>
-    <App />
+    <I18nProvider>
+      <App />
+    </I18nProvider>
   </React.StrictMode>,
 );